% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybinny.R
\name{mybinny}
\alias{mybinny}
\title{mybin function for table}
\usage{
mybinny(iter, n, p)
}
\arguments{
\item{iter}{a value}

\item{n}{a value}

\item{p}{a value}
}
\value{
table of mybin
}
\description{
mybin function for table
}
\examples{
mybinny(iter=1, n=10, p=0.7)

}
