% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2 function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{the number of iterations}

\item{x}{a value}

\item{fun}{a type of functino to perform, such as mean or sd}

\item{alpha}{a probability for a confidence interval}

\item{cx}{a value}

\item{...}{possible extra addictoins}
}
\value{
a histogram of the Bootstrap sample statistics
}
\description{
this myboot2 function will display a histogram of Bootstrap sample statistics based on the users inputs
}
\examples{
myboot2(x=rnorm( 20,mean=3,sd=2))
}
