% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{myclt Function}
\usage{
myclt(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{the base value used for sum of uniforms}

\item{iter}{the number of iterations the function will perform}

\item{a}{the left endmost point}

\item{b}{the right endmost point}
}
\value{
the result of myclt which is a distribution of the sum of uniforms
}
\description{
This function will display a distribution of the sum of uniforms
}
\examples{
w=myclt(n=50,iter=10000,a=5,b=10) # D
}
