% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{mymaxlik function}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{val}

\item{x}{val}

\item{param}{val}

\item{...}{vals}
}
\value{
stuff
}
\description{
mymaxlik function
}
\examples{
mymaxlik(x=c(9,9,1,9,9,9),param=seq(0,1,length=1000),lfun=logbin,xlab=expression(pi),main="Binomial",cex.main=2)
}
