% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{mymlnorm}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{a sample vector}

\item{mu}{the mean vector}

\item{sig}{the standard deviation vector}

\item{...}{possible additions but not necessary such as lwd}
}
\value{
normal distribution based on inputs
}
\description{
Displays a normal distribution based on the inputs
}
\examples{
mymlnorm(x=c(5,7,7,8,10),mu=seq(5,10,length=1000),sig=seq(0.1,4,length=1000),lwd=2,labcex=1)
}
