% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{the mean!}

\item{sigma}{the standard deviation!}

\item{a}{calculates up to x=a}
}
\value{
a value
}
\description{
myncurve
}
\examples{
myncurve(mu=10, sigma=4, a=25)
}
