% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myquad.R
\name{myquad}
\alias{myquad}
\title{Quadratic Function}
\usage{
myquad(x)
}
\arguments{
\item{x}{a value for x that is inputted by user}
}
\value{
the result of the quadratic function based on the x input
}
\description{
This is a basic quadratic function that will perform x^2-2x+x
}
\examples{
myquad(x = 0)
}
