% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{mysample}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n, iter, time}{numeric values}
}
\value{
distributions
}
\description{
mysample
}
\examples{
mysample()
}
