% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Project1 ntickets Function}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats on a flight}

\item{gamma}{probability that a plane actually is overbooked}

\item{p}{probability of success that a person shows up}
}
\value{
two graphs and listed values of nc, nd, gamma, N, and p
}
\description{
This function is for project1 to calculate the statistics for
the number of tickets sold on a flight based on the provided inputs
}
\examples{
ntickets(N=400,gamma = 0.02, p = 0.95)
}
