% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybinny.R
\name{mybinny}
\alias{mybinny}
\title{mybin function for table}
\usage{
mybinny(iter, n, p)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{a size inputted by user}

\item{p}{a probability}
}
\value{
table of mybin
}
\description{
Displays a table of mybin based on the user inputting iteration, size of n, and probability
}
\examples{
mybinny(iter=1, n=10, p=0.7)

}
