% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2 function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{the number of iterations}

\item{x}{a value}

\item{fun}{a value}

\item{alpha}{a value}

\item{cx}{a value}

\item{...}{a value}
}
\value{
a graph
}
\description{
myboot2 function
}
\examples{
myboot2(x=rnorm( 20,mean=3,sd=2))
}
