% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{myclt Function}
\usage{
myclt(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{a value}

\item{iter}{a value}

\item{a}{a value}

\item{b}{a value}
}
\value{
the result of myclt
}
\description{
myclt Function
}
\examples{
w=myclt(n=50,iter=10000,a=5,b=10) # D
}
