% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{mymaxlik function}
\usage{
mymaxlik(x)
}
\arguments{
\item{x}{a value}
}
\value{
value
}
\description{
Originally was the mymaxlik function, decided to go with another function to
read into R so this is just a placeholder for if I had used it
}
\examples{
mymaxlik(x=3)
}
