% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{the mean}

\item{sigma}{the standard deviation}

\item{a}{calculates up to x=a}
}
\value{
a normal distribution based on inputs
}
\description{
Provides a visualization of a normal distribution based on input mean, sigma, and a
}
\examples{
myncurve(mu=10, sigma=4, a=25)
}
